Move validate_user_settings into routes/aux
