+++ fractal_server/app/routes/aux/validate_user_settings.py
@@ -1,0 +1,41 @@
+from fastapi import HTTPException
+from fastapi import status
+from pydantic import ValidationError
+
+from fractal_server.app.db import AsyncSession
+from fractal_server.app.models import UserOAuth
+from fractal_server.app.models import UserSettings
+from fractal_server.app.routes.api.v2._aux_functions import logger
+from fractal_server.user_settings import SlurmSshUserSettings
+
+
+async def validate_user_settings(
+    *, user: UserOAuth, backend: str, db: AsyncSession
+):
+    """
+    FIXME docstring
+    """
+    # First: check that the foreign-key exists. TODO: remove this check,
+    # after this column is made required
+    if user.user_settings_id is None:
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail=f"Error: user '{user.email}' has no settings.",
+        )
+
+    user_settings = await db.get(UserSettings, user.user_settings_id)
+
+    if backend == "slurm_ssh":
+        try:
+            SlurmSshUserSettings(**user_settings.model_dump())
+        except ValidationError as e:
+            error_msg = (
+                "User settings are not valid for "
+                f"FRACTAL_RUNNER_BACKEND='{backend}'. "
+                f"Original error: {str(e)}"
+            )
+            logger.warning(error_msg)
+            raise HTTPException(
+                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+                detail=error_msg,
+            )
